pyproject.toml
[tool.poetry]
name = "devops-container-image-code-generator"
version = "1.20240710.1"
description = "GenAI Application which uses Langchain and source code repository files, such as dependency manifests, to generate container image code like Dockerfile and entrypoint script."
authors = ["Sagar Velankar <sagar.velankar@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/devopscodegen/devops-container-image-code-generator"
repository = "https://github.com/devopscodegen/devops-container-image-code-generator"
documentation = "https://github.com/devopscodegen/devops-container-image-code-generator"
keywords = ["devops", "container-image-code-generator", "devops-platform", "langchain", "genai"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Code Generators",
]
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
langchain-core = ">=0.1.5,<0.3"
langchain-openai = ">=0.0.1"
opentelemetry-distro = "^0.46b0"
opentelemetry-exporter-otlp = "^1.25.0"
opentelemetry-instrumentation-asyncio = "^0.46b0"
opentelemetry-instrumentation-aws-lambda = "^0.46b0"
opentelemetry-instrumentation-dbapi = "^0.46b0"
opentelemetry-instrumentation-logging = "^0.46b0"
opentelemetry-instrumentation-sqlite3 = "^0.46b0"
opentelemetry-instrumentation-threading = "^0.46b0"
opentelemetry-instrumentation-urllib = "^0.46b0"
opentelemetry-instrumentation-wsgi = "^0.46b0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-instrumentation-grpc = "^0.46b0"
opentelemetry-instrumentation-httpx = "^0.46b0"
opentelemetry-instrumentation-requests = "^0.46b0"
opentelemetry-instrumentation-tortoiseorm = "^0.46b0"
opentelemetry-instrumentation-urllib3 = "^0.46b0"
devopscodegen-common = "^1.20240710.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pytest = "^8.2.2"
black = "^24.4.2"
pylint = "^3.2.3"
langchain-cli = "^0.0.25"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
